[Style] book 관련 sidebar 적용
