+++ src/main/resources/templates/book/admin/viewAllBooks.html
@@ -1,0 +1,191 @@
+<!DOCTYPE html>
+<html lang="ko" xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>흑백문고 - 도서 목록</title>
+    <link rel="stylesheet" href="/css/style.css">
+    <link rel="stylesheet" href="/css/sidebar.css">
+    <style>
+        /* 테이블 기본 스타일 */
+        .book-table {
+            width: 100%;
+            border-collapse: collapse; /* 셀 간격 제거 */
+            margin-top: 20px;
+            background-color: #fff;
+            font-size: 0.9em;
+            box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1); /* 테이블 그림자 */
+        }
+
+        .book-table thead th {
+            background-color: #f5f5f5; /* 헤더 배경 */
+            color: #333;
+            font-weight: bold;
+            padding: 12px;
+            text-align: center;
+            border-bottom: 2px solid #ddd; /* 헤더 아래 테두리 */
+        }
+
+        .book-table tbody td {
+            padding: 12px;
+            text-align: center;
+            border-bottom: 1px solid #eee; /* 셀 사이 테두리 */
+            vertical-align: middle; /* 내용 중간 정렬 */
+        }
+
+        /* 행 교차 색상 */
+        .book-table tbody tr:nth-child(odd) {
+            background-color: #f9f9f9; /* 밝은 회색 */
+        }
+
+        .book-table tbody tr:nth-child(even) {
+            background-color: #fff; /* 흰색 */
+        }
+
+        /* 행 호버 효과 */
+        .book-table tbody tr:hover {
+            background-color: #f1f1f1; /* 마우스를 올렸을 때 강조 */
+            cursor: pointer;
+        }
+
+        /* 썸네일 이미지 */
+        .book-thumbnail {
+            width: 50px;
+            height: 70px;
+            object-fit: cover;
+            border-radius: 4px;
+        }
+
+        /* 버튼 스타일 */
+        .action-buttons {
+            display: flex; /* 버튼을 가로로 나열 */
+            gap: 8px; /* 버튼 간격 */
+            justify-content: center; /* 버튼 중앙 정렬 */
+        }
+
+        .action-buttons .button {
+            display: inline-block;
+            padding: 8px 10px; /* 버튼 안쪽 여백 */
+            font-size: 0.85em; /* 버튼 글꼴 크기 */
+            border-radius: 4px;
+            text-align: center;
+            text-decoration: none;
+            border: none;
+            cursor: pointer;
+            transition: background-color 0.3s ease, transform 0.2s ease;
+        }
+
+        /* 수정 버튼 스타일 */
+        .action-buttons .button.edit {
+            background-color: #555;
+            color: #fff;
+        }
+
+        .action-buttons .button.edit:hover {
+            background-color: #777;
+            transform: scale(1.01); /* 확대 효과 */
+        }
+
+        /* 삭제 버튼 스타일 */
+        .action-buttons .button.delete {
+            background-color: #d9534f; /* 빨간색 */
+            color: #fff;
+        }
+
+        .action-buttons .button.delete:hover {
+            background-color: #c9302c;
+            transform: scale(1.01); /* 확대 효과 */
+        }
+    </style>
+</head>
+<body>
+
+<!-- 헤더 불러오기 -->
+<header th:replace="~{admin/layout :: header}"></header>
+
+<!-- 메인 콘텐츠 -->
+<main class="admin-page-container">
+
+    <!-- 사이드바 -->
+    <aside id="admin-sidebar" th:replace="~{admin/admin-sidebar :: admin-sidebar}"></aside>
+
+    <!-- 오른쪽 콘텐츠 영역 -->
+    <section id="content-area">
+        <h3 class="content-title">도서 목록</h3>
+        <p class="content-description">등록된 도서 목록입니다.</p>
+
+        <!-- 도서 목록 테이블 -->
+        <section class="results-section" th:if="${books != null}">
+            <table class="book-table">
+                <thead>
+                <tr>
+                    <th>이미지</th>
+                    <th>제목</th>
+                    <th>저자</th>
+                    <th>카테고리</th>
+                    <th>출판사</th>
+                    <th>정가</th>
+                    <th>상태</th>
+                    <th>관리</th>
+                </tr>
+                </thead>
+                <tbody>
+                <tr th:each="book : ${books.content}" th:onclick="|location.href='/admin/books/' + ${book.id}|">
+                    <td><img th:src="${book.thumbnailImageUrl}" alt="도서 표지" class="book-thumbnail"></td>
+                    <td th:text="${book.title}">도서 제목</td>
+                    <td th:text="${#strings.arrayJoin(book.authors, ', ')}">저자</td>
+                    <td th:text="${#strings.arrayJoin(book.categories, ', ')}">카테고리</td>
+                    <td th:text="${book.publisher}">출판사</td>
+                    <td th:text="${book.standardPrice} + '원'">정가</td>
+                    <td th:switch="${book.bookStatus}">
+                        <span th:case="'IN_STOCK'">재고 있음</span>
+                        <span th:case="'OUT_OF_STOCK'">품절</span>
+                        <span th:case="'DISCONTINUED'">절판</span>
+                        <span th:case="*">알 수 없음</span>
+                    </td>
+                    <td>
+                        <div class="action-buttons">
+                            <a th:href="@{/admin/books/{bookId}/edit(bookId=${book.id})}" class="button edit">수정</a>
+                            <form th:action="@{/admin/books/{bookId}(bookId=${book.id})}" method="post" class="button-form delete-form">
+                                <input type="hidden" name="_method" value="DELETE" />
+                                <button type="submit" class="button delete">삭제</button>
+                            </form>
+                        </div>
+                    </td>
+                </tr>
+                </tbody>
+            </table>
+
+            <!-- 페이지네이션 컨트롤 -->
+            <div th:replace="~{fragments/pagination :: pagination(${books})}"></div>
+        </section>
+    </section>
+</main>
+
+<!-- 푸터 불러오기 -->
+<footer th:replace="~{admin/layout :: footer}"></footer>
+
+<script>
+    document.addEventListener("DOMContentLoaded", function () {
+        const deleteForms = document.querySelectorAll(".delete-form");
+
+        deleteForms.forEach(function (form) {
+            // 삭제 버튼 클릭 이벤트 처리
+            form.addEventListener("submit", function (event) {
+                event.preventDefault();
+
+                if (confirm("정말 삭제하시겠습니까?")) {
+                    form.submit();
+                }
+            });
+
+            const deleteButton = form.querySelector(".button.delete");
+            deleteButton.addEventListener("click", function (event) {
+                event.stopPropagation();
+            });
+        });
+    });
+
+</script>
+</body>
+</html>
